Add circularity metric and configurable cell property extraction (#18)
* Update mask_processing.py

- [draft] add circularity as extra regionprops property

* Update mask_processing.py

- refactor to distinguish cell property names from the properties (names: values) themselves
- add property: centroid

* Update mask_processing.py

- add doscstring for circularity

* Update mask_processing.py

- add centroids_yx attribute

* Make 64 bit integer type explicit

* Typing changes for Int64Array

- downstream refactoring changes

+++ src/arcadia_microscopy_tools/model.py
@@ -6,7 +6,7 @@ import numpy as np
 import torch
 from cellpose.models import CellposeModel
 
-from .typing import FloatArray, IntArray
+from .typing import FloatArray, Int64Array
 
 logger = logging.getLogger(__name__)
 
@@ -107,7 +107,7 @@ class SegmentationModel:
         batch_intensities: list[FloatArray],
         batch_size: int = 8,
         **cellpose_kwargs: dict[str, Any],
-    ) -> IntArray:
+    ) -> Int64Array:
         """Run cell segmentation using Cellpose.
 
         Args:
@@ -120,7 +120,7 @@ class SegmentationModel:
                 Common options include 'min_size' (minimum cell size in pixels).
 
         Returns:
-            IntArray: Output batch of labeled masks with shape (batch, height, width).
+            Int64Array: Output batch of labeled masks with shape (batch, height, width).
                 Each segmented cell has a unique positive integer ID, background is 0.
 
         Raises:
